+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### New
+- Fix darklist bypass []()
 - Add PEG token [#2002](https://github.com/kvhnuke/etherwallet/pull/2002)
 - Add Opiria PData custom gas [#2004](https://github.com/kvhnuke/etherwallet/pull/2004)
 - Add PASS, REDC, FTXT, 3LT, WMK, CRMT, ORCA and remove FXC [#2006](https://github.com/kvhnuke/etherwallet/pull/2006)